test(web): align composer E2E with input capability
